+++ packages/widgetbook/lib/metadata.dart
@@ -1,2 +1,2 @@
 /// The version as in pubspec.yaml
-const kWidgetbookVersion = '4.0.0-alpha.4';
+const kWidgetbookVersion = '4.0.0-alpha.5';