+++ packages/widgetbook/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- **BREAKING**: Merge `widgetbook_generator` and `widgetbook_test` packages into `widgetbook` package. ([#1772](https://github.com/widgetbook/widgetbook/pull/1772))
 - **FEAT**: Add `Defaults` class for default `setup` and `builder`. ([#1768](https://github.com/widgetbook/widgetbook/pull/1768))
 
 ## 4.0.0-alpha.4