+++ packages/widgetbook/CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 4.0.0-alpha.5
 
 - **BREAKING**: Merge `widgetbook_generator` and `widgetbook_test` packages into `widgetbook` package. ([#1772](https://github.com/widgetbook/widgetbook/pull/1772))
 - **REFACTOR**: Change default font to `Roboto`. ([#1773](https://github.com/widgetbook/widgetbook/pull/1773))